init implementation of rl wrapper
add a simple interface (AgentBase) to be used in different
time-evolution codes.

add SarsaAgent which inherits from AgentBase. It implements a simple
one-step action-value learning scheme. It takes a template parameter
"Model" which is the approximation type (for the q function).

add linear model, which has an independent linear model for each action,
and can do gradient descent update.